control structure

+++ 0x00-hello_world/6-division.py
@@ -3,6 +3,6 @@
 
 #division
 print(50/2)
-num1 = int(input("enter any number of your choice: "))
-num2 = int(input("enter any number of your choice: "))
+num1 = float(input("enter any number of your choice: "))
+num2 = float(input("enter any number of your choice: "))
 print(num1/num2)
